Lay widgets out in integer device pixels
The toolkit laid out in logical points and reconciled with the pixel
grid by multiplying by dpr, rounding, and dividing back -- Kit::snap,
snap_size and snap_rect, called from nearly every measure() and
arrange().  That bought a class of rounding errors rather than removing
one: positions rounded to nearest while sizes rounded up, so insetting a
rect grew it back out (hence the workaround in focus_ring); neighbours
snapped independently and disagreed, most visibly across the toolbar's
three slots; and grow-shares divided in floats lost up to a pixel per
child with no way to recover it.  The draw list then snapped everything
again on emit, so none of it was what made drawing crisp.

Widget geometry is now integral device pixels, and the arithmetic is
exact by construction.  Rect holds ints, the three snap primitives are
gone along with the round-versus-ceil asymmetry, and text metrics answer
in pixels directly rather than dividing by dpr for a caller that only
multiplied it back.  Growing children now share the remainder a pixel at
a time, so they fill their container exactly.

Design constants stay in points -- 6pt of padding has to stay 6pt on any
display -- and convert on use through Kit::px().  Converting rather than
baking keeps a sum of point terms rounding once instead of once per
term, and covers the per-widget pad_x/pad_y fields too.  Persistent
geometry stays in points for the same reason: dpr is per-window and
changes when a window moves between displays, so a stored pixel count
would change physical size.  That is why Page::sidebar_w is still a
point value.

Genuinely continuous quantities stay float and round only where they
become geometry: scroll offsets, the viewer's pan, zoom and rotation,
and pointer positions.  The scroll thumb keeps its ratios so its length
does not jitter mid-drag.

Points meet pixels at three boundaries now, each converting once: input
events and the host size on the way in, and the input method's caret
rectangle, the window-menu position and the resize mask on the way out.
The draw list is fed device pixels with a scale of 1.

+++ dn/kit-browser.hpp
@@ -136,7 +136,7 @@ struct Browser : Widget {
 	Browser(Kit &kit, Thumbnailer &thumbnailer);
 	~Browser() override;
 
-	void measure(Kit &kit, float max_w, float max_h) override;
+	void measure(Kit &kit, int max_w, int max_h) override;
 	void arrange(Kit &kit, Rect alloc) override;
 	void paint(Kit &kit) const override;
 	void prepare(Kit &kit) override;
